Use custom pads for jumper footprints (#716)
In KiCad 4 there were no polygon pads which resulted in some overlapping
pad constructs. This is no longer required, as well as the old system
causes some nasty problems like duplication of pin numbers and
unconnected warnings. This PR solves the issue for the jumpers by
converting the pads to custom shaped pads as appropriate.