Adjust dimensions slightly so the resulting card isn't too tight.
The official specs at
http://www.s100computers.com/General%20Images/IEEE_696_1983.pdf clearly
states that the edge should be 6.350 inches to 6.360 inches to fit in a
slot measuring 6.375 inches.  That's easy to miss though.  According to
this data sheet (http://www.edac.net/dat/files/161.pdf) for slots
matching the spec, the slot is 6.375 inches wide and the card edge
should be 0.012 inches smaller.  EDAC makes slots of varying connector
counts matching the S100 pitch for applications other than S100.