Add scripted TI S-PWSON-N8 (replaces #541)
This adds a footprint for TI S-PWSON-N8. The footprint uses normal qfn
ipc settings (nearest result to suggested footprint). The suggested
footprint is not reached as it seems to be quite old (TI suggests the
long obsolete IPC-SM-782. telling me that datasheet is quite old or at
least the suggested footprint is.)

The EP size is inclded as we did for similar footprints in the QFN
library. (same custom format string as ti qfn footprints)

There is one more thermal via as the script does not support the
generation of the exact pattern suggested by TI. (The drill size is a
bit reduced so it should balance out.)

Datasheet link: http://www.ti.com/lit/ds/symlink/lp2951.pdf (drawing on
page 27)

script PR: https://github.com/pointhi/kicad-footprint-generator/pull/244