Replace existiong LQFP 32 and 36 with scripted variants
This replaces the 32 and 36 pin versions of the LQFP footprints with
scripted ones.

LQFP-32-1EP_5x5mm_P0.5mm_EP3.45x3.45mm is deleted as that one
is kind of a frankenstein created from another qfp footprint with the
thermal pad definition of a qfn footprint. (Not used by any symbol in
the current lib.)