Replace existiong LQFP 32 and 36 with scripted variants (#1176)
* Replace existiong LQFP 32 and 36 with scripted variants

This replaces the 32 and 36 pin versions of the LQFP footprints with
scripted ones.

LQFP-32-1EP_5x5mm_P0.5mm_EP3.45x3.45mm is deleted as that one
is kind of a frankenstein created from another qfp footprint with the
thermal pad definition of a qfn footprint. (Not used by any symbol in
the current lib.)

* lqfp 36 7x7 with small pitch ipc definition

Changed this 0.65mm pitch device to use the <=0.625mm pitch ipc rules
to get enough clearance between the pads.