Update 16-pin Textool socket to match all the others.
I originally submitted the 16-pin Textool socket in PR #465 as a test
case of my script.  Once it was accepted, I submitted the rest of the
Textool sockets in PR #495.  However, a bunch of additional
requirements were places on the second PR, so the new sockets ended up
with a different naming convention, and some different dimensions,
than the 16-pin socket.

This PR brings the 16-pin socket in line with all the others.