NKK GW12LxH SPDT right-angle rocker switch (#1279)
http://www.nkkswitches.com/pdf/GW.pdf

The datasheet specifies the pins as 4, 5, and 6, but I numbered them
1, 2, and 3, to allow the footprint to be used with the existing
library symbol for an SPDT switch.  The alternative would be to create
a new library symbol where the pins are numbered 4, 5, and 6.

I left the mounting pins unnumbered.  The alternative would be to
number them MP.  This would then require a new library symbol, too,
since there does not seem to be a symbol for a switch with a mounting
pin.