Remove dublicate TI VQFN-RGR-20
This footprint is called S-PVQFN-N20 in the datasheet assigned to it.
The assigned datasheet does not specify the EP size as it is the general
datasheet for all packages with this name.
http://www.ti.com/lit/ml/mpqf239/mpqf239.pdf

The 2x2mm ep size of the footprint seems to be wrongly derived.

This footprint is not used by any symbol right now so no change is
needed on the footprint side of things.

We have two S-PVQFN-N20 footprints in the lib (2.4x2.4mm ep and
2.7x2.7mm ep)