Replace TI VQFN-RGW-20 with scripted version
This footprint is called S-PVQFN-N20 in the datasheet assigned to it.
The assigned datasheet does not specify the EP size as it is the general
datasheet for all packages with this name.
http://www.ti.com/lit/ml/mpqf122b/mpqf122b.pdf

But the footprint is used in the library for TPS7A7200RGW
Its datasheet is http://www.ti.com/lit/ds/symlink/tps7a7200.pdf

The new footprint assigns the correct name to it.