Create footprint for C&K switch PTS125SM43SMTR2-LFS (#1495)
* Create footprint for C&K switch PTS125SM43SMTR2-LFS

This is a footprint for PTS125 series SMT switches from C&K. For example, this footprint is good for PTS125SM43SMTR2-LFS particular switch.

* Trying to solve travis indicated issues

* Adding tags

* Fix Fabrication Layer related errors on travis

   Fabrication layer requirements
    No drawings found on fabrication layer
    Second Reference Designator missing
     - Add RefDes to F.Fab layer with '%R'

* Fixing Travis error

Violating F5.2
    Fabrication layer requirements
    No drawings found on fabrication layer

* Fixing Travis error

  Violating F9.3
    3D model settings
    3D model file path missing from the 3D model settings of the footprint

* Changes in F. CrtYard, F. Silks

Adopted good things from existing footprints of PTS645 and PTS810, so as to have a common design language.

* Fix Travis Error; Improve F. Silks

* Rename the footprint based on switch naming convention

* Remove file with old footprint name

* Added PSMTR at the end of current file name.

* Delete SW_Push_1P1T_NO_CK_PTS125.kicad_mod

Delete file with old name.

* Silk screen width changed to 0.12

* one silk line segment width was 0.15. Fixed it now.

* Looks like even though silk screen width standard is 0.12, the reference label of silk screen should have width of 0.15. Fixed that now.

* Again forgot to change width of one line segment of silk screen to 0.12. Now fixed at the source too.

* Center Ref Des for the F. Fab layer

* Changed file name as well as the footprint name

* Delete file with old name

As the footprint name has changed, the file name also need to change. This is taken care by deleting the file and adding new file with new file name. As part of last commit, I already added the new file, so now deleting the old file with this commit.