Diode_SMD:/D_MELF-RM10_Universal: Reduce exposed copper on back
This footprint exposed rectangular pads on both front and copper layers.
When soldering SMD diodes, a large amount of solder can be wicked to the
back pads. This is not ideal, as a mountain of solder accumulates on the
back pads.
To resolve this, only expose a circular pad on the back.