Fix hole sizes in 2.54 mm headers. (#2178)
* Fix hole sizes in 2.54 mm headers.

The holes were previously 0.8 mm diameter, which is too small to
accommodate a standard header.  Normal header posts are 0.025" square
(0.635 mm), and so the diagonal (approx. 0.898 mm) will not fit into
a 0.8 mm hole.  1.0 mm is an appropriate size.

* Fix hole sizes in 2.54 mm headers (for Arduino Nano footprints).

The same modification made in the parent commit:

    The holes were previously 0.8 mm diameter, which is too small to
    accommodate a standard header.  Normal header posts are 0.025" square
    (0.635 mm), and so the diagonal (approx. 0.898 mm) will not fit into
    a 0.8 mm hole.  1.0 mm is an appropriate size.

now applied to the Nano footprints.