extend selection of DIN41612 connectors (#1076)
* extend selection of DIN41612 connectors

This selection was generated by an updated script using the
kicad-footprint-generator. This current selection was built by checking
the catalogs of ept, ERNI and Harting and collect the most standard
series.
Added in this commit were: B (was already present), C, D, E, F, Q, R
Still missing standard series are:
  - H11, H15 (vertical and horizontal are different from each other and
    require an updated script, only H11 horizontal is present)
  - M (has a lot of variations which needs to be tracked).
There are also many manufacturer specific variants which might follow
(some are very easy to add to the script, others not so much).

The new script was made to copy the style of the existing scripts, just
adding the increased flexiblity to generate all series. Changes to the
generated footprints compared to the existing B series are:
- a1 pin is a roundrect instead of rect
- vertical connectors got their silk ref moved to the side. They are
  often used on backplanes, so there might be no space on top.
- courtyard of the horizontal connector is simpler on the connector
  side. This space if off the PCB anyways and there needs to be room to
  attach the connector.
- names are now always using two digit numbers for rows and pins.
- Q and R series vertical have slightly more drawings, trying to
  symbolise their inverted connector nature (vertical is male, while
  usually they are female).

There are two series where the names contain the rows which are
occupied. These are due to connectors only ept sells (E and F series
2x16). The standard is to leave the center row empty, ept also sells a
variant where the top row is empty. To distinguish those only the
non-standard has the rows added to the name.

As a general remark to these connectors. The naming of the pins and rows
is very strict, based on the location relative to the mounting position.
At least in vertical this always holds true, horizontal got some
flexibility on the spacing. The origin is always located at the a1 pin,
even if it is not present. This ensures that connectors can be switched
without having to change everything. This was already the case for the
existing connectors.

* update autogenerated DIN41612 connectors:

this is the output of
scripts/Connector/Connector_IEC_DIN/generate_din41612.py  in
kicad-footprint-generator at revision
e610537351f77d457bc2f90f021258384de20fc8

* update DIN41612 connectors from updated script

* fix colliding silk in DIN41612 F vertical

* remove WIP H15 connectors

added them by accident, not done yet

* din41612: update script generated connectors

* din41612: regenerate connectors from script

* din41612: update script output

* din41612: increase copper on large holes

* din41612: update from script output

- increases silk screen distance
- fixes pin numbering

* din41612: updated with increased clearance

* din41612: update (horiz pin 1 moved & larger clearance)

* integrate review feedback