Fix pin numbering in 1.5mm pitch Pico-Lock connectors (504050 series)
The existing footprints have reversed pin ordering, so that pad 1 does
not match that on the datasheet or indicated by the pin 1 indicator on
the connector itself.

The footprints are created with kicad-footprint-generator with

```
smd_single_row_plus_mounting_pad.py conn_molex.yaml
```

Footprints generated by this script must always have pin 1 on the
left-most side, so in order to fix the numbering the connectors are
rotated by 180 degrees.

This update also fixes incorrect tags that specified "top entry" instead
of "horizontal".