CUI PD-30S connector. (#592)
* CUI PD-30S connector.

https://www.cui.com/product/resource/digikeypdf/pd-30s.pdf

This is a shielded version of the PD-30 connector which is already in
the library.  The PD-30S can use the same schematic symbol as the
PD-30.

I've provided the PD-30S in versions with and without circular holes.

* PD-30S: Update description and add line on fab layer.

As discussed in the PR:

* Added "DIN" to the description.
* Added a line to the fab layer to distinguish the round part from the
  square part.

* CUI_PD-30S: rename new shield pins to SH, as discussed in #592.

* CUI_PD-30S: Surprise!  The datasheet has changed.

This PR was originally submitted 2+ years ago based on this datasheet:

https://github.com/KiCad/kicad-footprints/files/5271623/pd-30s.pdf

However, this month a new datasheet was released, with many of the
dimensions changed by 0.1mm:

https://www.cui.com/product/resource/pd-30s.pdf

Fun!