More feathers (#2507)
* Fix mounting holes on Adafruit_Feather

Turns out that the upper mounting holes are 2.5mm, not 2.54mm on all
Feather boards. The bottom mounting holes are usually 2.54mm, but
without copper, so use that. Some Feathers have copper on the botom
mounting holes, or slightly smaller (2.2mm) holes, but that should
probably still be ok.

* Update documentation url for Adafruit_Feather_WithMountingHoles

This was already updated for the version without mounting holes, but
forgotten here.

* Add variants of the Adafruit_Feather footprint

This adds footprints for all Feather modules that use a slightly
different footprint (extra pins or longer board).

* Renumber extra pins to be counterclockwise

* Update 3D models

* Round sizes to 0.01mm for Feather FONA footprint

* Fix mounting hole positions on M0 Wifi